# docs/api_contracts/public-proposals.openapi.yaml
openapi: 3.1.0
info:
  title: Vitarerum — Public Proposal Submission API
  version: 1.0.0
  description: |
    Unauthenticated endpoints that let **any citizen** open a collection-use proposal
    without an account. They back the public frontend at `/submit-proposal`
    (Angular feature `src/app/features/public/`).

    ## Trust model — read this first
    Every field in these requests is **untrusted input from the open internet**.
    The frontend's client-side validation, the Cloudflare Turnstile token, and the
    honeypot field are **deterrents only**. The actual protection MUST be enforced
    here, on the server:

    - **Verify the Turnstile token** server-side via Cloudflare `siteverify`
      (the secret key never leaves the server). A client token alone proves nothing.
    - **Rate-limit** per source IP, per e-mail, and globally → respond `429` with
      `Retry-After`.
    - **Sanitize & bound every input**: enforce the documented `maxLength`s, strip
      control characters, store as plain text, and **escape on render** in the staff
      UI (defend against stored XSS). Reject CR/LF in any field used to build an
      e-mail (header/-injection defence).
    - **Double opt-in**: `POST /public/proposals` creates only a *pending, unverified*
      record and e-mails a single-use, signed, expiring token. The proposal is
      materialised into the staff queue **only** on `POST /public/proposals/confirm`.
      Unconfirmed records are discarded after the TTL.
    - **Honeypot**: if `website` is non-empty, respond `202` as if successful but
      perform **no work** (accept-and-drop) so bots learn nothing.
    - **No file uploads** on these endpoints. Attachments only happen later, inside
      the authenticated conversation thread.
    - **No credentials**: these endpoints MUST NOT accept or rely on cookies/auth
      headers. Lock CORS to the public origin; front with a WAF / bot-management.

    ## Relationship to the authenticated API
    These are distinct from the authenticated `POST /proposals` (which requires a
    session + `X-Permission-Id`). A public submission has no account, so it captures
    the citizen's own name + e-mail as the requester contact.

servers:
  - url: http://127.0.0.1:8000/api/v1
    description: Local development
  - url: https://api.vitarerum.example/api/v1
    description: Production (place behind WAF / Cloudflare)

tags:
  - name: Public proposals
    description: Unauthenticated citizen submission flow (double opt-in).

paths:
  /public/proposals:
    post:
      tags: [Public proposals]
      operationId: submitPublicProposal
      summary: Open a pending proposal submission (step 1 of double opt-in)
      description: |
        Validates the payload, verifies the Turnstile token, applies rate limits,
        and — if everything passes — creates a **pending, unverified** submission and
        e-mails a confirmation link to `citizenEmail`. No proposal is visible to staff
        until the citizen confirms.

        Honeypot: a non-empty `website` yields a `202` with no side effects.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublicProposalSubmission'
            examples:
              citizen:
                summary: A student requesting access
                value:
                  citizenName: Pedro Silva
                  citizenEmail: pedro@example.test
                  subject: Acesso à Coleção de Zoologia
                  body: Gostaria de estudar um espécime para a minha tese de mestrado.
                  consent: true
                  captchaToken: 0.AbC...turnstile-response-token
                  website: ''
      responses:
        '202':
          description: |
            Accepted. A confirmation e-mail has been (or, for honeypot hits, appears to
            have been) sent. The proposal does not exist yet.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSubmissionReceipt'
              example:
                status: PENDING_CONFIRMATION
                email: pedro@example.test
        '422':
          description: |
            Validation failed (missing/invalid fields, consent not given). Request
            validation is normalised app-wide to 422; the field errors are carried
            under `errors`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServerError' }
              example:
                message: Validation failed
                errors:
                  - field: citizenEmail
                    message: A valid e-mail address is required.
        '403':
          description: |
            Turnstile verification failed (missing, invalid, or expired token).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServerError' }
              example:
                message: Captcha verification failed.
        '429':
          description: Rate limit exceeded for this IP/e-mail.
          headers:
            Retry-After:
              description: Seconds to wait before retrying.
              schema: { type: integer, example: 60 }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServerError' }
              example:
                message: Too many requests. Please try again later.
        '503':
          description: Captcha provider unreachable; submission could not be verified.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServerError' }

  /public/proposals/confirm:
    post:
      tags: [Public proposals]
      operationId: confirmPublicProposal
      summary: Confirm a submission via the e-mailed token (step 2 of double opt-in)
      description: |
        Consumes the single-use token from the confirmation link. On success the
        pending submission is materialised into the staff proposal queue. The token
        is invalidated after use (idempotency: a second call returns
        `ALREADY_CONFIRMED`).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublicConfirmationRequest'
            example:
              token: eyJhbGciOiJIUzI1NiIs...single-use-signed-token
      responses:
        '200':
          description: |
            The token was processed. Inspect `status` for the outcome — note that
            `EXPIRED` and `INVALID` are returned as `200` (not errors) so the public
            page can render a friendly message. Use a non-2xx only for unexpected
            server faults.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicConfirmationResult'
              examples:
                confirmed:
                  value: { status: CONFIRMED, referenceNumber: VRP-20260625-0007 }
                alreadyConfirmed:
                  value: { status: ALREADY_CONFIRMED, referenceNumber: VRP-20260625-0007 }
                expired:
                  value: { status: EXPIRED }
                invalid:
                  value: { status: INVALID }
        '429':
          description: Rate limit exceeded.
          headers:
            Retry-After:
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServerError' }

components:
  schemas:
    PublicProposalSubmission:
      type: object
      additionalProperties: false
      required: [citizenName, citizenEmail, subject, body, consent, captchaToken]
      properties:
        citizenName:
          type: string
          minLength: 1
          maxLength: 120
          description: Full name of the requesting citizen.
        citizenEmail:
          type: string
          format: email
          maxLength: 180
          description: Contact e-mail; the confirmation link is sent here.
        subject:
          type: string
          minLength: 1
          maxLength: 160
        body:
          type: string
          minLength: 1
          maxLength: 4000
          description: The opening message describing the request.
        consent:
          type: boolean
          const: true
          description: RGPD consent. MUST be `true`; reject otherwise.
        captchaToken:
          type: string
          minLength: 1
          description: |
            Cloudflare Turnstile response token. The server MUST verify it via
            `siteverify`; never trust it as-is.
        website:
          type: string
          maxLength: 255
          default: ''
          description: |
            **Honeypot.** Hidden from real users; should be empty. A non-empty value
            marks the request as a bot — the server accept-and-drops (`202`, no work).
            It is deliberately NOT schema-rejected (a `422` would reveal to the bot
            that the field is monitored), only bounded in length.

    PublicSubmissionReceipt:
      type: object
      additionalProperties: false
      required: [status, email]
      properties:
        status:
          type: string
          const: PENDING_CONFIRMATION
        email:
          type: string
          format: email
          description: Echoed back so the UI can show where the link was sent.

    PublicConfirmationRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
          minLength: 1
          description: Single-use, signed, expiring token from the e-mailed link.

    PublicConfirmationResult:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [CONFIRMED, ALREADY_CONFIRMED, EXPIRED, INVALID]
        referenceNumber:
          type: string
          description: Present once a proposal exists (CONFIRMED/ALREADY_CONFIRMED).
          example: VRP-20260625-0007

    ServerError:
      type: object
      description: |
        Error body shape consumed by the frontend `toApiError()`
        (`src/app/core/http/api-error.model.ts`). It reads `message` and the
        field-error array. Request-validation failures (`422`) carry that array
        under `errors`; `toApiError()` reads both `errors` and `fieldErrors`.
      properties:
        message:
          type: [string, 'null']
          description: Human-readable summary; surfaced to the user.
        errors:
          type: array
          description: Field errors from request validation (`422`).
          items:
            $ref: '#/components/schemas/ValidationFieldError'
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationFieldError'
      additionalProperties: true

    ValidationFieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          example: citizenEmail
        message:
          type: string
          example: A valid e-mail address is required.